core/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==23.2.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2025.3.0
gast==0.4.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.7
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kucoin-python==1.0.23
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.10.1
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
oauthlib==3.2.2
openai==1.65.5
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
rubicon-objc==0.5.0
safetensors==0.5.3
scikit-base==0.12.0
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.23.1
setuptools==76.0.0
six==1.16.0
sktime==0.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
starlette==0.46.1
sympy==1.13.1
termcolor==2.4.0
tesseract==0.1.3
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.2
uvicorn==0.34.0
websocket-client==1.8.0
websockets==12.0
wheel==0.45.1
wrapt==1.14.1
wsproto==1.2.0
